+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.0] - 2021-07-29
+
+### Added
+
+- Support for multiple to-addresses
+- Support cc-addresses
+- Support for bcc-addresses
+
+### Fixed
+
+- Improves verbose output
+
 ## [0.2.3] - 2021-04-08
 
 ### Fixed